updated 23

+++ admin/functions.php
@@ -58,7 +58,7 @@ function deleteCategories()
 function findAllPost()
 {
     global $conn;
-    $query = "SELECT * FROM post";
+    $query = "SELECT * FROM post ORDER BY post_id DESC";
     // $query = "SELECT * FROM category LIMIT 5";
     $select_all_post_sidebar = mysqli_query($conn, $query);
 
@@ -72,6 +72,7 @@ function findAllPost()
         $post_tags = $row['post_tags'];
         $post_comment_count = $row['post_comment_count'];
         $post_date = $row['post_date'];
+        $post_view_count = $row['post_view_count'];
         echo "<tr>";
 ?>
         <td><input type='checkbox' class='checkBoxes' name='checkBoxArray[]' value='<?php echo $post_id ?>'></td>
@@ -92,9 +93,10 @@ function findAllPost()
         echo "<td>{$post_tags}</td>";
         echo "<td>{$post_comment_count}</td>";
         echo "<td>{$post_date}</td>";
+        echo "<td>{$post_view_count}  <a href='posts.php?reset={$post_id}'>Reset</a></td>";
         echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
         echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are you sure? you want to delete');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
         echo "</tr>";
     }
 }
@@ -106,6 +108,17 @@ function deletePost()
         $the_post_id = $_GET['delete'];
         $query = "DELETE FROM post WHERE post_id = {$the_post_id} ";
         $delete_query = mysqli_query($conn, $query);
+        header("Location: posts.php");
+    }
+}
+
+function resetPost()
+{
+    global $conn;
+    if (isset($_GET['reset'])) {
+        $the_post_id = $_GET['reset'];
+        $query = "UPDATE post SET post_view_count = 0 WHERE post_id = {$the_post_id} ";
+        $reset_query = mysqli_query($conn, $query);
         header("Location: posts.php");
     }
 }
